requirements.txt
absl==0.0
addict==2.4.0
anthropic==0.67.0
anytree==2.13.0
apex==0.9.10dev
AppKit==0.2.8
astor==0.8.1
azureml==0.2.7
backoff==2.2.1
beartype==0.21.0
beautifulsoup4==4.13.5
borb==3.0.2
boto3==1.40.30
botocore==1.34.69
chardet==4.0.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.12
dashscope==1.24.4
decord==0.6.0
deepspeed==0.17.5
dm_env==1.6
docx==0.2.4
easyocr==1.7.2
einops==0.8.1
ewmh==0.1.6
fastapi==0.116.1
fastchat==0.1.0
fastdtw==0.3.4
filelock==3.19.1
fitz==0.0.1.dev2
flash_attn==2.8.3
Flask==3.1.2
formulas==1.3.0
Foundation==0.1.0a0.dev1
func_timeout==4.3.5
fuzzywuzzy==0.18.0
groq==0.31.1
grpc==1.0.0
grpc_tools==1.0.0
gymnasium==1.2.0
httpx==0.28.1
imagehash==4.3.2
imageio==2.33.1
immutabledict==4.2.1
ipdb==0.13.13
ipython==8.12.3
jax==0.7.1
jsonlines==4.0.0
jsonschema==4.25.1
librosa==0.11.0
liger_kernel==0.6.2
LnkParse3==1.5.2
loguru==0.7.3
lxml==5.2.1
matplotlib==3.10.6
mutagen==1.47.0
numpy==2.3.3
oa_atomacos==3.2.0
odfpy==1.4.1
openai==1.107.2
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
paramiko==4.0.0
pdfplumber==0.11.7
peft==0.17.1
Pillow==11.3.0
playwright==1.55.0
psutil==5.9.0
py7zr==1.0.0
pyautogui==0.9.54
pycocotools==2.0.10
pydantic==2.11.9
pydrive==1.3.1
pydub==0.25.1
pygetwindow==0.0.9
pypdf==6.0.0
PyPDF2==3.0.1
pyperclip==1.9.0
pyshine==0.0.9
pytesseract==0.3.13
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
python_Levenshtein==0.27.1
pytz==2024.1
pywinauto==0.6.9
PyYAML==6.0.1
PyYAML==6.0.2
Quartz==0.0.1.dev0
qwen_vl_utils==0.0.11
rapidfuzz==3.14.1
Requests==2.32.5
requests_toolbelt==1.0.0
safetensors==0.6.2
scipy==1.16.2
screeninfo==0.8.1
Send2Trash==1.8.2
sentencepiece==0.2.1
setuptools==75.1.0
skimage==0.0
supervision==0.26.1
tenacity==8.2.3
termcolor==3.1.0
tiktoken==0.11.0
timeout_decorator==0.5.0
timm==1.0.19
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.5
transformers==4.56.1
uiautomator2==3.4.1
ultralytics==8.3.199
uvicorn==0.35.0
win32gui==221.6
win32security==2.1.0
wrapt_timeout_decorator==1.5.1
Xlib==0.21
xmltodict==1.0.0
yapf==0.40.2
zhipuai==2.1.5.20250825
